+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # CHANGELOG for PHP CS Fixer: custom fixers
 
 ## v3.15.0
+- Deprecate DataProviderNameFixer - use "php_unit_data_provider_name"
 - Deprecate PhpdocParamOrderFixer
-- Update minimum PHP CS Fixer version to 3.17.0
+- Update minimum PHP CS Fixer version to 3.19.0
 
 ## v3.14.0
 - Add EmptyFunctionBodyFixer